refactor: modularize backend routes and services
